requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
contextlib2==21.6.0
Django==5.0.7
django-ninja==1.2.2
django-ninja-extra==0.21.2
django-ninja-jwt==5.3.2
ninja-schema==0.13.6
injector==0.22.0
pydantic==2.8.2
pydantic_core==2.20.1
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
django-encrypted-model-fields==0.6.5
psycopg2==2.9.9
python-dotenv==1.0.1
google-generativeai